api/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
ansible==2.9.13           # via -r requirements.in
appnope==0.1.0            # via ipython
asgiref==3.2.10           # via django
attrs==20.2.0             # via pytest
backcall==0.2.0           # via ipython
certifi==2020.6.20        # via requests
cffi==1.14.2              # via cryptography
chardet==3.0.4            # via requests
click==7.1.2              # via pip-tools
cryptography==3.1         # via ansible
decorator==4.4.2          # via ipython
django-debug-toolbar==2.2  # via -r requirements.in
django-environ==0.4.5     # via -r requirements.in
django-extensions==3.0.9  # via -r requirements.in
django==3.0.5             # via -r requirements.in, django-debug-toolbar, djangorestframework
djangorestframework==3.11.1  # via -r requirements.in
gunicorn==20.0.4          # via -r requirements.in
idna==2.10                # via requests
iniconfig==1.0.1          # via pytest
invoke==1.4.1             # via -r requirements.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.18.1           # via -r requirements.in
isort==5.5.2              # via -r requirements.in
jedi==0.17.2              # via ipython
jinja2==2.11.2            # via ansible
markupsafe==1.1.1         # via jinja2
more-itertools==8.5.0     # via pytest
packaging==20.4           # via pytest
parso==0.7.1              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==5.3.1          # via -r requirements.in
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.7     # via ipython
psycopg2-binary==2.8.6    # via -r requirements.in
ptyprocess==0.6.0         # via pexpect
py==1.9.0                 # via pytest
pycparser==2.20           # via cffi
pygments==2.7.1           # via ipython
pyparsing==2.4.7          # via packaging
pytest-django==3.10.0     # via -r requirements.in
pytest==6.0.2             # via -r requirements.in, pytest-django
pytz==2020.1              # via django
pyyaml==5.3.1             # via ansible
requests==2.24.0          # via -r requirements.in
six==1.15.0               # via cryptography, packaging, pip-tools
sqlparse==0.3.1           # via django, django-debug-toolbar
toml==0.10.1              # via pytest
traitlets==5.0.4          # via ipython
urllib3==1.25.10          # via requests
wcwidth==0.2.5            # via prompt-toolkit
APScheduler==3.6.3
Pyrogram==0.17.1
TgCrypto==1.2.0
celery==4.4.1
redis==3.5.3
django-redis==4.12.1
django-celery-beat
django-celery-results
django-modeltranslation==0.15.1
# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
